Manage listings

+++ app/Http/Controllers/ListingController.php
@@ -49,6 +49,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!');
@@ -63,6 +65,11 @@ class ListingController extends Controller
     // Update listing data from form
     public function update(Request $request, Listing $listing)
     {
+        // Is authorized?
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -85,8 +92,25 @@ class ListingController extends Controller
     // Delete listing
     public function destroy(Listing $listing)
     {
+        // Is authorized?
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action');
+        }
+
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    // Show manage listings
+    public function manage()
+    {
+        // return view('listings.manage');
+        return view('listings.manage', [
+            'listings' => auth()
+                ->user()
+                ->listings()
+                ->get()
+        ]);
     }
 }
